website/requirements.txt
aniso8601==9.0.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Limiter==2.0.2
Flask-RESTful==0.3.9
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
limits==2.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
typing_extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
gunicorn
